refactor(runtimed): move request dispatcher boundary
