+++ crates/runtimed/src/notebook_sync_server/metadata.rs
@@ -3671,121 +3671,6 @@ pub(crate) async fn update_kernel_presence(
     }
 }
 
-/// Short label for a request variant, used in telemetry logs.
-///
-/// Returns a static string — no allocation — suitable for structured logging
-/// fields.  Intentionally avoids `Debug` formatting because some variants
-/// carry large payloads (doc bytes, comm messages) that would bloat log lines.
-pub(crate) fn request_label(req: &NotebookRequest) -> &'static str {
-    match req {
-        NotebookRequest::LaunchKernel { .. } => "LaunchKernel",
-        NotebookRequest::ExecuteCell { .. } => "ExecuteCell",
-        NotebookRequest::ExecuteCellGuarded { .. } => "ExecuteCellGuarded",
-        NotebookRequest::ClearOutputs { .. } => "ClearOutputs",
-        NotebookRequest::InterruptExecution { .. } => "InterruptExecution",
-        NotebookRequest::ShutdownKernel { .. } => "ShutdownKernel",
-        NotebookRequest::RunAllCells { .. } => "RunAllCells",
-        NotebookRequest::RunAllCellsGuarded { .. } => "RunAllCellsGuarded",
-        NotebookRequest::SendComm { .. } => "SendComm",
-        NotebookRequest::GetHistory { .. } => "GetHistory",
-        NotebookRequest::Complete { .. } => "Complete",
-        NotebookRequest::SaveNotebook { .. } => "SaveNotebook",
-        NotebookRequest::CloneAsEphemeral { .. } => "CloneAsEphemeral",
-        NotebookRequest::SyncEnvironment { .. } => "SyncEnvironment",
-        NotebookRequest::ApproveTrust { .. } => "ApproveTrust",
-        NotebookRequest::ApproveProjectEnvironment { .. } => "ApproveProjectEnvironment",
-        NotebookRequest::GetDocBytes { .. } => "GetDocBytes",
-    }
-}
-
-/// Handle a NotebookRequest and return a NotebookResponse.
-pub(crate) async fn handle_notebook_request(
-    room: &Arc<NotebookRoom>,
-    request: NotebookRequest,
-    daemon: std::sync::Arc<crate::daemon::Daemon>,
-) -> NotebookResponse {
-    debug!(
-        "[notebook-sync] Handling request: {}",
-        request_label(&request)
-    );
-
-    match request {
-        NotebookRequest::LaunchKernel {
-            kernel_type,
-            env_source,
-            notebook_path,
-        } => {
-            crate::requests::launch_kernel::handle(
-                room,
-                &daemon,
-                kernel_type,
-                env_source,
-                notebook_path,
-            )
-            .await
-        }
-
-        NotebookRequest::ExecuteCell { cell_id } => {
-            crate::requests::execute_cell::handle(room, cell_id).await
-        }
-
-        NotebookRequest::ExecuteCellGuarded {
-            cell_id,
-            observed_heads,
-        } => crate::requests::execute_cell::handle_guarded(room, cell_id, observed_heads).await,
-
-        NotebookRequest::ClearOutputs { cell_id } => {
-            crate::requests::clear_outputs::handle(room, cell_id).await
-        }
-
-        NotebookRequest::InterruptExecution {} => {
-            crate::requests::interrupt_execution::handle(room).await
-        }
-
-        NotebookRequest::ShutdownKernel {} => crate::requests::shutdown_kernel::handle(room).await,
-
-        NotebookRequest::RunAllCells {} => crate::requests::run_all_cells::handle(room).await,
-
-        NotebookRequest::RunAllCellsGuarded { observed_heads } => {
-            crate::requests::run_all_cells::handle_guarded(room, observed_heads).await
-        }
-
-        NotebookRequest::SendComm { message } => {
-            crate::requests::send_comm::handle(room, message).await
-        }
-
-        NotebookRequest::GetHistory { pattern, n, unique } => {
-            crate::requests::get_history::handle(room, pattern, n, unique).await
-        }
-
-        NotebookRequest::Complete { code, cursor_pos } => {
-            crate::requests::complete::handle(room, code, cursor_pos).await
-        }
-
-        NotebookRequest::SaveNotebook { format_cells, path } => {
-            crate::requests::save_notebook::handle(room, &daemon, format_cells, path).await
-        }
-
-        NotebookRequest::CloneAsEphemeral { source_notebook_id } => {
-            crate::requests::clone_notebook::handle(&daemon, source_notebook_id).await
-        }
-
-        NotebookRequest::SyncEnvironment { guard } => {
-            crate::requests::sync_environment::handle(room, guard).await
-        }
-
-        NotebookRequest::ApproveTrust {
-            dependency_fingerprint,
-        } => crate::requests::approve_trust::handle(room, dependency_fingerprint).await,
-
-        NotebookRequest::ApproveProjectEnvironment { project_file_path } => {
-            crate::requests::approve_project_environment::handle(room, project_file_path).await
-        }
-
-        NotebookRequest::GetDocBytes {} => crate::requests::get_doc_bytes::handle(room).await,
-    }
-}
-
 /// Promote inline deps from CRDT metadata to a project file.
 ///
 /// When a kernel is project-backed (pixi:toml, uv:pyproject, or
